+++ doc/CHANGELOG.md
@@ -30,6 +30,7 @@ See doc/COPYRIGHT.rdoc for more details.
 # Changelog
 
 * `#2618` Fix: When issues are renamed to work packages all watcher assignments are lost
+* `#2623` [Journals] Images in journal notes are not displayed
 * Improved newline handling for journals
 
 ## 3.0.0pre24